Split Noise handshake tests out into a separate class
